requirements.txt
Django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.21.5
requests==2.31.0
python-telegram-bot==21.7
mysqlclient==2.1.1
virtualenv==20.26.4
SQLAlchemy==2.0.35
python-dotenv==1.0.0
setuptools==68.0.0
Pillow==10.0.0
aiohttp==3.8.5
qrcode[pil]==7.4.2
emoji==2.8.0
crcmod==1.7
coverage==7.2.7
pytest==7.4.2
pytest-django==4.5.2